Made the app prompt the user for their name if it's detected that it's their first time opening the app.
Created a 'name-prompt' activity that is called when it's detected that there is no device id stored in the preferences file.
On submitting the name, a new profile document is created in firebase with the name of the user and the device id. The document's name is the device id.
The device id is then stored in the preference file of the device.