Refactor Event and Profile classes, introduce EventTable class
This commit refactors the Event and Profile classes to extend from the DocumentSerializable class for better management of Firebase data. It modifies EventPosterImageHandler to use the new EventTable class for database operations instead of directly sending event data. The HomeFragment and UploadPosterActivity classes have also been updated to use EventTable for CRUD operations on event data.