Remove dashboard feature and rename QR code scanner activity
Deleted all files related to the dashboard feature from the application including its layout, navigation and fragment files. Additionally, renamed QRcodeScanner class to QRCodeScannerActivity to maintain the naming conventions for activity classes in the project. These changes are part of an effort to simplify the application and enhance code readability.