Add ProfileImageGenerator with related tests
Added a new file, ProfileImageGenerator, in the profile package with the function to generate a profile image based on the name. Also, added a new file, ProfileImageGeneratorTest, in the test package to validate the correctness of the profile image generation. It checks both the size of the generated images and the extraction of initials for use in the image.