Add default user profile image and loading functionality
Implemented a feature that generates and sets a default user profile picture upon profile creation. Also, updated the loading screen activity to fetch and set this profile image when loading the user's profile. This introduces a visual element to user profiles and enhances functionality by properly loading the profile image on startup.