Show profile image on MainActivity
Introduced function to create a circular bitmap for profile images in ProfileImageManager. Extended user profile to check if a bitmap is set and consequently updated MainActivity to initialize a circular profile image at startup. This allows for a more appealing aesthetic and a smoother user experience by providing a visual element for every profile early in the process.