Add cancel button to profile creation page
A cancel button has been added to the profile creation page in the user interface. The behavior associated with this button has been defined in the ProfileEditActivity.java. Upon clicking this button, the user is directed away from the current activity. This update improves user experience by providing a way to exit out of profile creation without making changes.