Add QR code generation to event creation process
Implemented the QR code generation process at the point of creating an event in the EnterEventDetailsActivity. Two QR codes - for "checkIn" and "eventInfo" - are now generated for each event when it's not in editing mode. Additionally, moved the QRCodeManager class to the QR code package, allowing a more organized code structure.