event check-in and update profile process
Updated the check-in function in the Event class to use 'replace' instead of 'put' to update the attendees info. This makes sure that the existing attendee's information is properly updated. Also, enhanced the updateProfile function in the SuccessfulCheckinActivity class by ensuring the id of the joined event is added to the user profile. This makes the tracking of user participation in events more precise.