Now gets the location on the startup of SuccessfulCheckinActivity
If location access has already previously been granted, the user's location coordinates are now gotten on the startup of SuccessfulCheckinActivity.
Previously if the user had already granted access to their location and tried to check in with their location, there location would not been gotten and the checkin coordinates would be zero.