Update event details and sign-up process
Revised how attendees sign up for or express interest in an event: the process is now managed within the EventDetailsDialogFragment view. Also removed the EventSignUpActivity class and adjusted related classes accordingly. The UI for event details has been enhanced for better user experience with an inclusion of 'I'm interested' button.