Update UI and add new profile icon in the navigation menu
The user interface of the event details, profile page, and event explore page are updated. The profile circle icon in the bottom navigation menu is replaced with the baseline account circle icon. In the profile page, a LinearLayout is implemented to better organize elements and modifications to margins, fonts, and styles are applied. Additionally, unnecessary padding from the event explore page is removed.