Refactor ImageView cropping and dimension setting
The image cropping method in ImageViewHelper.java was refactored to handle images with different orientations properly. It also adjusts the dimensions of ImageView based on the aspect ratio of the image, whether it is horizontal or vertical. Concurrently, all calls of the old method "cropPosterToImage" have been replaced by calling the adjusted method "cropImageToAspectRatio".