Refactor profile fragment and address empty fields
The previous 'ProfileUpdatedFragment' has been renamed to 'ProfileViewFragment' to reflect its purpose more accurately. In addition to this, elements within the profile layout including TextView Id and handling of empty contact and homepage fields have been improved. The code will now hide the separator when the contact or homepage information is not available to enhance the user interface.